refactor(DrumGridDark): extract keyboard editor hook and sub-components
Move all keyboard navigation state and handlers (cursor position, variation
menu, message auto-clear, cell ref management, handleKeyboardEditKeyDown)
from DrumGridDark.tsx into a new useKeyboardEditor hook.

DrumGridDark: 789 → 448 lines. New hook: src/hooks/useKeyboardEditor.ts (423 lines).
All 239 tests pass. No behavior changes.
